modified app to render Greeting component
