+++ src/App.js
@@ -1,28 +1,16 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { getGreeting } from './redux/app/app';
+import React from 'react';
+import {
+  BrowserRouter,
+} from 'react-router-dom';
+import Greeting from './components/Greeting';
 
 function App() {
-  const dispatch = useDispatch();
-  const fetchApi = 'http://localhost:3000/api/v1/greetings';
-
-  const myGreeting = useSelector((state) => state.greeting);
-
-  useEffect(() => {
-    const apiGreeting = async () => {
-      const fetchGreeting = await fetch(fetchApi);
-      const greeting = await fetchGreeting.json();
-      return dispatch(getGreeting(greeting));
-    };
-    if (myGreeting.length === 0) {
-      apiGreeting();
-    }
-  }, []);
   return (
-    <div>
-      <h1>Hello react front end</h1>
-      <h2>{myGreeting.greeting}</h2>
-    </div>
+    <>
+      <BrowserRouter>
+        <Greeting />
+      </BrowserRouter>
+    </>
   );
 }
